# openapi.yaml
openapi: 3.1.0
info:
  title: S.I.R.K. HTTP API
  description: HTTP interface for the S.I.R.K. coding-agent service. The service has no authentication or TLS. Bind it only to a trusted network. Paths supplied in `directory` identify directories as seen by the host running S.I.R.K.; they are not resolved relative to the client. Requests to an unknown path return `404` with a JSON error. A method other than the documented method returns `405` with a JSON error.
  version: 0.1.0
servers:
- url: http://127.0.0.1:8080
  description: Default address used by `sirk http`.
paths:
  /health:
    get:
      summary: Check whether the service is available
      operationId: health
      responses:
        '200':
          description: The service accepted the health-check request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
  /openapi.yaml:
    get:
      summary: Download the OpenAPI specification
      description: Returns the complete specification served by this S.I.R.K. binary as plain UTF-8 text.
      operationId: getOpenApiDocument
      responses:
        '200':
          description: The OpenAPI document.
          content:
            text/plain:
              schema:
                type: string
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
  /swagger:
    get:
      summary: Open the Swagger UI
      description: Returns an HTML page that loads Swagger UI and configures it to display `/openapi.yaml`.
      operationId: getSwaggerUi
      responses:
        '200':
          description: The Swagger UI HTML page.
          content:
            text/html:
              schema:
                type: string
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
  /v1/agent/run:
    post:
      summary: Run a named agent
      description: Loads the agent definition from `.agents/<agent>.md` below the supplied directory and runs it with the literal `input` string. Template-like content in `input`, including `{{ ... }}`, is passed through unchanged. The result is the agent's final response after any enabled tool requests have been handled. `X-Sirk-Flow-Id` is required and must identify a flow created for the same directory; the complete prompt and each model response are appended to that flow's transcript.
      operationId: runAgent
      parameters:
      - name: X-Sirk-Flow-Id
        in: header
        description: Required identifier returned by POST /v1/flows for this directory.
        required: true
        schema:
          type: string
        example: flow-18f-1234-0
      requestBody:
        description: Agent execution request.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
            example:
              agent: code-explainer
              directory: /workspace/project
              input: Explain src/lib.rs.
        required: true
      responses:
        '200':
          description: The agent completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
              example:
                result: The module exposes the public S.I.R.K. API.
        '400':
          description: The request body or X-Sirk-Flow-Id header is invalid, incomplete, unknown, or does not belong to the supplied directory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Missing or invalid X-Sirk-Flow-Id header
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
        '500':
          description: The requested agent operation could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: An operation-specific error message.
  /v1/flows:
    post:
      summary: Create a flow transcript
      description: Creates one empty transcript in `directory` and returns its identifier. Include the returned value as the required `X-Sirk-Flow-Id` header on every later agent, TREE, and Git request for this execution. Agent calls append each complete prompt delivered to the model and each complete model response to this single transcript. The transcript contains no agent metadata, resumable state, or tool-only records.
      operationId: createFlow
      requestBody:
        description: Server-visible directory where the flow transcript is stored.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectoryRequest'
            example:
              directory: /workspace/project
        required: true
      responses:
        '201':
          description: The flow transcript was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowResponse'
              example:
                flowId: flow-18f-1234-0
        '400':
          description: The request body is invalid, incomplete, or contains an unknown field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid request
        '500':
          description: The flow transcript could not be created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: An operation-specific error message.
  /v1/git/add:
    post:
      summary: Stage all changes below a directory
      description: Runs Git add for every change below `directory`, including deletions. The requested directory must exist and be within a Git working tree visible to the server. `X-Sirk-Flow-Id` is required and must identify a flow created for the same directory.
      operationId: gitAdd
      parameters:
      - name: X-Sirk-Flow-Id
        in: header
        description: Required identifier returned by POST /v1/flows for this directory.
        required: true
        schema:
          type: string
        example: flow-18f-1234-0
      requestBody:
        description: Server-visible Git directory.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectoryRequest'
            example:
              directory: /workspace/project
        required: true
      responses:
        '200':
          description: All changes below the requested directory were staged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: ok
        '400':
          description: The request body or X-Sirk-Flow-Id header is invalid, incomplete, unknown, or does not belong to the supplied directory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Missing or invalid X-Sirk-Flow-Id header
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
        '500':
          description: The Git add operation could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: An operation-specific error message.
  /v1/git/status:
    post:
      summary: List changed paths below a directory
      description: Returns sorted, unique relative UTF-8 paths for modified, untracked, and deleted regular files in the Git project below `directory`. Paths hidden by `.treeignore` are excluded. The requested directory must exist and be within a Git working tree visible to the server. `X-Sirk-Flow-Id` is required and must identify a flow created for the same directory.
      operationId: gitStatus
      parameters:
      - name: X-Sirk-Flow-Id
        in: header
        description: Required identifier returned by POST /v1/flows for this directory.
        required: true
        schema:
          type: string
        example: flow-18f-1234-0
      requestBody:
        description: Server-visible Git directory.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectoryRequest'
            example:
              directory: /workspace/project
        required: true
      responses:
        '200':
          description: Changed paths were listed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatusResponse'
              example:
                paths:
                - src/lib.rs
                - README.md
        '400':
          description: The request body or X-Sirk-Flow-Id header is invalid, incomplete, unknown, or does not belong to the supplied directory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Missing or invalid X-Sirk-Flow-Id header
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
        '500':
          description: The Git status operation could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: An operation-specific error message.
  /v1/tree:
    post:
      summary: List files below a directory
      description: Returns sorted, unique relative UTF-8 paths for tracked and non-ignored untracked files below `directory`. Paths hidden by `.treeignore` are excluded. The requested directory must exist and be within a Git working tree visible to the server. `X-Sirk-Flow-Id` is required and must identify a flow created for the same directory.
      operationId: tree
      parameters:
      - name: X-Sirk-Flow-Id
        in: header
        description: Required identifier returned by POST /v1/flows for this directory.
        required: true
        schema:
          type: string
        example: flow-18f-1234-0
      requestBody:
        description: Server-visible Git directory.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectoryRequest'
            example:
              directory: /workspace/project
        required: true
      responses:
        '200':
          description: Files were listed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
              example:
                paths:
                - README.md
                - src/lib.rs
        '400':
          description: The request body or X-Sirk-Flow-Id header is invalid, incomplete, unknown, or does not belong to the supplied directory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Missing or invalid X-Sirk-Flow-Id header
        '405':
          description: The endpoint does not accept the HTTP method used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Method not allowed
        '500':
          description: The TREE operation could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: An operation-specific error message.
components:
  schemas:
    AgentRunRequest:
      type: object
      required:
      - directory
      - agent
      - input
      properties:
        agent:
          type: string
          description: Identifier of the `.agents/<agent>.md` definition to run.
          examples:
          - code-explainer
        directory:
          type: string
          description: Server-visible execution directory.
          examples:
          - /workspace/project
        input:
          type: string
          description: Literal text passed to the agent without template expansion.
          examples:
          - Explain src/lib.rs.
      additionalProperties: false
    AgentRunResponse:
      type: object
      required:
      - result
      properties:
        result:
          type: string
          description: Final text returned by the executed agent.
      additionalProperties: false
    DirectoryRequest:
      type: object
      required:
      - directory
      properties:
        directory:
          type: string
          description: Server-visible execution directory.
          examples:
          - /workspace/project
      additionalProperties: false
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable failure message.
      additionalProperties: false
    FlowResponse:
      type: object
      required:
      - flowId
      properties:
        flowId:
          type: string
          description: Identifier required in X-Sirk-Flow-Id for the flow's later requests.
      additionalProperties: false
    GitStatusResponse:
      type: object
      required:
      - paths
      properties:
        paths:
          type: array
          items:
            type: string
          description: Sorted, unique paths relative to the requested directory.
      additionalProperties: false
    HealthResponse:
      type: object
      required:
      - status
      properties:
        status:
          $ref: '#/components/schemas/SuccessStatus'
          description: Fixed success indicator.
      additionalProperties: false
    StatusResponse:
      type: object
      required:
      - status
      properties:
        status:
          $ref: '#/components/schemas/SuccessStatus'
          description: Fixed success indicator.
      additionalProperties: false
    SuccessStatus:
      type: string
      enum:
      - ok
    TreeResponse:
      type: object
      required:
      - paths
      properties:
        paths:
          type: array
          items:
            type: string
          description: Sorted, unique paths relative to the requested directory.
      additionalProperties: false